Rearrange and rename a few things in the $lib folder
